requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.3
pillow==10.2.0
portablemc==4.1.0
PyOpenGL==3.1.7
pyopengltk==0.0.4
requests==2.31.0
urllib3==2.1.0
